requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.7
django-admin-bootstrapped==2.4.0
django-embed-video==1.1.2
django-modeladmin-reorder==0.2
easy-thumbnails==2.4.1
idna==2.5
MySQL-python==1.2.5
olefile==0.44
Pillow==4.2.1
requests==2.18.3
sorl-thumbnail==12.3
urllib3==1.22
wheel==0.24.0
